drm: don't light up unlit monitors
If a monitor is dark when plymouth is started, we shouldn't
try to light it up.  There are rules on which outputs can be
attached to which controllers and we're very likely to break
those rules just assigning things willy-nilly.

The kernel should set us up in the way it thinks is best anyway.
There's no reason to second guess the kernel here.